Reported changed attributes correctly
When an attribute gets updated because a height, it will also update the whole row, so the heights are aligned.
When doing so, the attributes that were updated, were not properly registered, hence never filled into the `invalidatedItemIndexPaths` array. This caused the collectionView NOT to update that attribute.

This is fixed by registering those attributes as well

Fixes #17