Fixed invalidateRepeatCounts
This issue came along when invalidating a repeat count of a section with multiple bricks. The logic assumed that the repeat count was set for simular bricks and only for appending. This fix will now check if the attributes have the same identifiers. And if not, the invalidation of the cells will start with the first attribute that has a different identifier

Fixes #101