Support for ordering bricks within a section
Added support to specify which brick is repeated at what index for a section.

Say a section has 2 bricks, that are each repeated 5 times. If the `orderDataSource` is set, the implementor needs to specify based on the actual index (IndexPath’s item), which brick and index needs to be returned. This gives full control to the implementor to define the order of the bricks

Fixes #139